requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=1.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
moviepy>=1.0.3
pillow>=10.0.0
requests==2.31.0
google-cloud-texttospeech>=2.14.1
google-cloud-vision>=3.4.4
pytest>=8.0.2
pytest-asyncio==0.23.5
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0
flask==3.0.2
ollama==0.1.6
anthropic>=0.3.11
google-generativeai>=0.1.1
cohere>=4.37
huggingface-hub>=0.20.3
